Comment out the test theme 'greenclaire' by default. Update comment.
The 'greenclaire' entry in the list of themes is useful to validate the
'remote theme' feature, i.e., check that the server correctly handles
the requests to a theme that is hosted on a remote HTTP server.
'greenclaire' points to test files under on http://nano.local/ddd and
is therefore useable inside Taodyne only.