Add all variations for the Black and White theme.
All page templates for the theme 'Black and White' have a drop-down list
that allows the user to make a selection among the various sub-themes that
are available (BlackOnWhite, WhiteOnBlack, and so on).

Implementation:
- The data model for each page template has an additional 'theme' field which
is the name to be used in the 'theme' statement in the .ddd file.
- The page property views use a 'mixin' to reach a common function that
returns the configuration info for the combobox. Base classes are modified
so that their initComponent() method accepts an array of item configurations,
which are inserted first when the list of items is created.
- On the server side, the generation function is given page.theme.

FIXME: I have noticed a message in the console: 'Unexpected: timer !== 0'
(from PageControllerBase.js). To reproduce: click on one page, change its
sub-theme, click on another page. No other visible side-effect.