Fixed issues with fields with direct value, refs #3374 @2.5h
This was a vexing issue in the way we are submitting fields, that was
actually causing several fields to be sent as blank instead of as their
proper value. It had been silent, I only saw it when it triggered
an exception in multiview_texture.

The new code ensures that if what we get after we remove labels and init
is empty (i.e. {}), this is not what we send as the initialization value
to the user interface (and then to the JSON file on disk).